Feat: core.common 모듈 build-logic 적용
