+++ core/common/src/main/kotlin/com/kolown/porring/core/common/component/CoilImage.kt
@@ -1,10 +1,8 @@
-package com.kolown.common.component
+package com.kolown.porring.core.common.component
 
-import android.util.Log
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
 import androidx.compose.foundation.combinedClickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -27,29 +25,30 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
-import com.kolown.common.R
 import com.kolown.designsystem.ui.theme.Primary
 import com.kolown.designsystem.ui.theme.PrimaryUnActive
 import com.kolown.designsystem.ui.theme.Surface2
+import com.kolown.porring.core.common.R
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
-fun NoRippleCoilImage(
+fun CoilImage(
     modifier: Modifier = Modifier,
     onClick: () -> Unit = {},
     onLongClick: () -> Unit = {},
-    onDoubleClick: () -> Unit = {},
     delay: Long = 0L,
     imageUrl: String,
     isTextExist: Boolean = true,
+    onDoubleClick: () -> Unit = {},
+    onClickEnabled: Boolean = true
 ) {
-    val interactionSource = remember { MutableInteractionSource() }
     var isLoading by remember { mutableStateOf(true) }
     var isError by remember { mutableStateOf(false) }
     val coroutineScope = rememberCoroutineScope()
     val loadingModifier = modifier.shimmerEffect()
+
 
     if (isError) {
         Column(
@@ -78,19 +77,18 @@ fun NoRippleCoilImage(
             }
         }
     } else {
+
         Box(
             modifier = if (isLoading) loadingModifier else modifier,
         ) {
             AsyncImage(
-                modifier = Modifier
+                modifier = if (onClickEnabled) Modifier
                     .fillMaxSize()
                     .combinedClickable(
-                        indication = null,
-                        interactionSource = interactionSource,
                         onClick = onClick,
                         onLongClick = onLongClick,
                         onDoubleClick = onDoubleClick
-                    ),
+                    ) else Modifier.fillMaxSize(),
                 model = imageUrl,
                 contentDescription = null,
                 contentScale = ContentScale.Crop,
@@ -110,7 +108,7 @@ fun NoRippleCoilImage(
                     isError = true
                 }
             )
+
         }
     }
-
 }
