+++ core/common/src/main/kotlin/com/kolown/porring/core/common/component/DetailTopAppBar.kt
@@ -1,4 +1,4 @@
-package com.kolown.common.component
+package com.kolown.porring.core.common.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -18,7 +18,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
-import com.kolown.common.R
+import com.kolown.porring.core.common.R
 
 @Composable
 fun DetailTopAppBar(
