+++ core/common/src/main/kotlin/com/kolown/porring/core/common/component/SnackBarBridge.kt
@@ -1,4 +1,4 @@
-package com.kolown.common.component
+package com.kolown.porring.core.common.component
 
 import com.kolown.model.SnackBarEvent
