+++ feature/home/src/main/java/com/kolown/porring/feature/home/component/RandomImageList.kt
@@ -37,9 +37,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.kolown.common.component.CoilImage
-import com.kolown.common.component.FollowDialog
-import com.kolown.common.component.LocalSnackBarBridge
+import com.kolown.porring.core.common.component.CoilImage
+import com.kolown.porring.core.common.component.FollowDialog
+import com.kolown.porring.core.common.component.LocalSnackBarBridge
 import com.kolown.designsystem.ui.theme.Primary
 import com.kolown.model.PostContentModel
 import com.kolown.model.Reactions
