+++ feature/main/src/main/java/com/kolown/porring/feature/main/MainScreen.kt
@@ -22,7 +22,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.lifecycleScope
-import com.kolown.common.component.showSnackBarWithData
+import com.kolown.porring.core.common.component.showSnackBarWithData
 import com.kolown.designsystem.ui.theme.PrimaryDark
 import com.kolown.designsystem.ui.theme.SnackBarContainer
 import com.kolown.porring.feature.main.component.MainBottomBar
