+++ feature/search/src/main/java/com/kolown/porring/feature/search/DetailSearchScreen.kt
@@ -21,9 +21,9 @@ import androidx.compose.ui.Modifier
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
-import com.kolown.common.component.DetailItem
+import com.kolown.porring.core.common.component.DetailItem
 import com.kolown.designsystem.ui.theme.PrimaryContainerDark
-import com.kolown.common.component.DetailTopAppBar
+import com.kolown.porring.core.common.component.DetailTopAppBar
 import com.kolown.model.PostContentModel
 import com.kolown.model.Reactions
 import kotlinx.coroutines.delay
