+++ feature/search/src/main/java/com/kolown/porring/feature/search/component/PostItem.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
-import com.kolown.common.component.CoilImage
+import com.kolown.porring.core.common.component.CoilImage
 import com.kolown.model.PostContentModel
 
 @Composable
